[Gold III] Title: 행성 X3, Time: 420 ms, Memory: 64248 KB -BaekjoonHub

+++ 백준/Gold/2830. 행성 X3/행성 X3.java
@@ -17,8 +17,8 @@ public class Main {
 
         long answer = 0L;
         for (int i = 0; i < DIGIT; i++) {
-            int zeroCount = residentsNum - oneCount[i];
-            answer +=   (long) zeroCount * oneCount[i] * (1L << i) ;
+            long zeroCount = residentsNum - oneCount[i];
+            answer += (zeroCount * oneCount[i] * (1L << i));
         }
         System.out.println(answer);
     }
